+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## Uncommitted
+### Changed
+- Default is not nucleotide, not amino acid
+
 ## [0.1.0] - 2019-05-22
 ### Added
 - ToDo list now available